+++ src/Tests/Unit/Public/Get-CFAccount.Tests.ps1
@@ -12,39 +12,35 @@ Import-Module $PathToManifest -Force
 #-------------------------------------------------------------------------
 
 InModuleScope 'pwshCloudflare' {
-    Describe 'Get-CFZone Function Tests' -Tag Unit {
+    Describe 'Get-CFAccount Function Tests' -Tag Unit {
         BeforeAll {
             $WarningPreference = 'SilentlyContinue'
             $ErrorActionPreference = 'SilentlyContinue'
             # Mock the dependent cmdlets and variables
-            Mock Invoke-CFRestMethod { return @{ result = @(@{ Name = 'TestZone'; Id = '12345' }) } }
             $script:cfBaseApiUrl = 'https://api.cloudflare.com/client/v4'
+            $script:cfAccountLookupTable = @{'myAccount' = '12345' }
             $script:cfSession = New-Object Microsoft.PowerShell.Commands.WebRequestSession
+            Mock Invoke-CFRestMethod { return [PSCustomObject]@{ result = [PSCustomObject]@{ AccountName = 'myAcct' } } }
         }
         # Context 'Error' {
         # }
         Context 'Success' {
-            It 'Calls Invoke-CFRestMethod with correct parameters for ZoneName' {
-                Get-CFZone -ZoneName 'example.com'
-                Assert-MockCalled Invoke-CFRestMethod -Exactly 1 -Scope It -ParameterFilter {
-                    $Uri -eq 'https://api.cloudflare.com/client/v4/zones?name=example.com' -and
-                    $Method -eq 'GET'
-                }
+            It 'Should get account by name' {
+                $Result = Get-CFAccount -AccountName 'myAcct'
+                $Result.AccountName | Should -Be 'myAcct'
             }
-            It 'Calls Invoke-CFRestMethod with correct parameters for ZoneID' {
-                Get-CFZone -ZoneID '12345'
-                Assert-MockCalled Invoke-CFRestMethod -Exactly 1 -Scope It -ParameterFilter {
-                    $Uri -eq 'https://api.cloudflare.com/client/v4/zones?id=12345' -and
-                    $Method -eq 'GET'
-                }
+            It 'Should get account by id' {
+                $Result = Get-CFAccount -AccountId '12345'
+                $Result.AccountName | Should -Be 'myAcct'
             }
-            It 'Returns objects of type Cloudflare.Zone' {
-                $result = Get-CFZone -ZoneName 'example.com'
-                $result.PSObject.TypeNames[0] | Should -BeExactly 'Cloudflare.Zone'
+            It 'Should get all accounts' {
+                $Result = Get-CFAccount
+                $Result.AccountName | Should -Be 'myAcct'
             }
-        }
-        AfterAll {
-            $script:cfSession | Remove-Variable
+            It 'Should return object of type Cloudflare.Account' {
+                $Result = Get-CFAccount
+                $result.PSObject.TypeNames[0] | Should -BeExactly 'Cloudflare.Account'
+            }
         }
     }
 }