+++ src/pwshCloudflare/Private/Invoke-CFRestMethod.ps1
@@ -29,9 +29,18 @@ function Invoke-CFRestMethod {
         [object]$Body
     )
     try {
+        if (-not $PSBoundParameters['WebSession']) {
+            $PSBoundParameters.Add('WebSession', $script:cfSession)
+        }
         Invoke-RestMethod @PSBoundParameters
     }
     catch {
+        if ($_.ErrorDetails.Message -eq 'page not found') {
+            # TODO: This could be that the resource doesn't exists, for example if the user input
+            # the wrong database name. Update this error handling to dump $PsBoundParameters to
+            # give the user more information about what went wrong.
+            throw $_
+        }
         $ApiError = $_.ErrorDetails.Message | ConvertFrom-Json
         if ($ApiError.errors.count -eq 1) {
             throw "Cloudflare API returned error code $($ApiError.errors.code) with message: $($ApiError.errors.message)"