+++ src/pwshCloudflare/Public/Get-CFZone.ps1
@@ -38,9 +38,6 @@ function Get-CFZone {
     begin {
         Write-Verbose "$($MyInvocation.MyCommand.Name) :: BEGIN :: $(Get-Date)"
         Write-Verbose "ParameterSetName: $($PSCmdlet.ParameterSetName)"
-        if (-not $script:cfSession) {
-            throw 'Cloudflare session not found. Use Set-CloudflareSession to create a session.'
-        }
     }
     process {
         if ($ZoneName) {