+++ src/pwshCloudflare/Public/New-CFZoneRecord.ps1
@@ -30,6 +30,7 @@
 #>
 function New-CFZoneRecord {
     [CmdletBinding()]
+    [OutputType('Cloudflare.ZoneRecord')]
     param(
         [Parameter(ParameterSetName = 'ZoneId', Mandatory)]
         [string]$ZoneId,
@@ -67,9 +68,6 @@ function New-CFZoneRecord {
     begin {
         Write-Verbose "$($MyInvocation.MyCommand.Name) :: BEGIN :: $(Get-Date)"
         Write-Verbose "ParameterSetName: $($PSCmdlet.ParameterSetName)"
-        if (-not $script:cfSession) {
-            throw 'Cloudflare session not found. Use Set-CloudflareSession to create a session.'
-        }
     }
     process {
         if ($ZoneName) {