+++ src/pwshCloudflare/Public/Remove-CFZoneRecord.ps1
@@ -31,9 +31,6 @@ function Remove-CFZoneRecord {
     begin {
         Write-Verbose "$($MyInvocation.MyCommand.Name) :: BEGIN :: $(Get-Date)"
         Write-Verbose "ParameterSetName: $($PSCmdlet.ParameterSetName)"
-        if (-not $script:cfSession) {
-            throw 'Cloudflare session not found. Use Set-CloudflareSession to create a session.'
-        }
     }
     process {
         if ($ZoneName) {